Amplitude_weekly report/requirements.txt
openai>=1.40.0
flask>=3.0.0
google-generativeai>=0.4.0